More correct web socket close Promise handling
A recent commit forced processing of messages sent to the web socket
handler to happen asynchronously, to avoid deadlocks when an `await`
of the message body would prevent incoming frames to be processed in
multi-frame messages. Unfortunately, that could cause the close
Promise to be kept before all message processing was done. This
change makes sure that won't happen.