Speed up payload masking up to 200x
The idiom used to perform frame payload masking turned out to be *extremely*
slow, utterly dominating the time to serialize or parse a masked frame.
Because all frames sent from the client to the server must be masked, this
cost in turn dominated round trip websocket performance.

On my machine the new (simpler!) idiom, which makes better use of native Blob
functionality, performs the masking itself up to 200x faster (for a 64KiB
payload).  This greatly reduces the penalty for using masking versus the
unmasked baseline (measured as total time across a mix of frames up to 64KiB):

    (Masking Penalty)   BEFORE   AFTER
    FrameParser           400x    2.3x
    FrameSerializer       400x    6.6x