Fix a WebSocket client lookup when pinged
If the server sent a ping while we were trying to send a message, it
could result in circular waiting, since:

1. We'd start sending a message, meaning the outgoing Supply is busy
2. The ping arrives while we're serializing the message, meaning the
   `supply` block in the TCP/TLS connector is busy with the `emit` of
   that received message, and so the send cannot happen.
3. The ping wants to send a pong, but its `emit` blocks, inducing a
   deadlock.

It's possible that the TCP/TLS connector should not try and enforce the
kind of concurrency control that it does so that this kind of problem
cannot happen in the first place, but that will need closer inspection.
In the meantime, sending the pong asynchronously is enough to resolve
the problem.